Port to udisks2
The upstream udisks project have changed how their command-line tool
works in quite a few ways:
 * `udisks` han renamed to `udisksctl`.
 * Command-line arguments have been changed.
 * Format of information output (`info` and `dump`) has been changed.
 * Removed some useful information from the output, such as model,
   vendor and whether a CD/DVD is blank. This is presumably only
   available via D-Bus or maybe it's been removed completely.
 * The command-line tool now provides `unlock` and `lock` options
   so these can be used in the default bashmount custom commands
   instead of calling `cryptsetup` directly. This is the only
   user-visible change to the upstream udisks project that actually
   benefits bashmount.

The changes to bashmount required are mainly when parsing information
about the device. The default options and configuration file also
needed to be reworked.

There is no longer any option to eject a device, though preumably it
is available via D-Bus. bashmount is now relying on the external
`eject` command.